+++ src/heart/device/selection.py
@@ -12,7 +12,7 @@ def select_device(*, x11_forward: bool) -> Device:
     layout_mode = Configuration.device_layout_mode()
     orientation: Orientation
     if layout_mode == DeviceLayoutMode.CUBE:
-        orientation: Orientation = Cube.sides()
+        orientation = Cube.sides()
     else:
         orientation = Rectangle.with_layout(
             columns=Configuration.device_layout_columns(),
