+++ src/heart/environment.py
@@ -2,4 +2,4 @@
 
 from heart import DeviceDisplayMode  # noqa: F401
 from heart.runtime.game_loop import GameLoop  # noqa: F401
-from heart.runtime.render_pipeline import RendererVariant  # noqa: F401
+from heart.runtime.rendering.variants import RendererVariant  # noqa: F401
