+++ src/heart/runtime/rendering/constants.py
@@ -1,0 +1,3 @@
+from typing import Literal
+
+RGBA_IMAGE_FORMAT: Literal["RGBA"] = "RGBA"
